+++ src/components/EntityInfoCard/EntityInfoCard.component.tsx
@@ -12,7 +12,7 @@ import {
 import { entityCardContainerStyles } from "./EntityInfoCard.styles";
 
 import { FlexColumnLayout } from "@layouts";
-import { useTestId, useMergedProps, useLogger } from "@utils";
+import { useTestId, useMergedProps } from "@utils";
 
 import type {
   EntityInfoCardProps,
@@ -31,7 +31,7 @@ const defaultEntityInfoCardProps: EntityInfoCardDefaultProps = {
 /**
  * Displays key information about an entity in a card.
  *
- * @version 0.0.12
+ * @version 0.0.13
  *
  * @param {EntityInfoCardProps & TestIdProps} props - EntityInfoCard component props
  * @returns {JSX.Element}
@@ -39,7 +39,6 @@ const defaultEntityInfoCardProps: EntityInfoCardDefaultProps = {
 const EntityInfoCard = (
   props: EntityInfoCardProps & TestIdProps
 ): JSX.Element => {
-  const logger = useLogger("EntityInfoCard");
   const mergedProps = useMergedProps(defaultEntityInfoCardProps, props);
   const { testId: testIdProp, overrideTestId } = props;
   const testId = useTestId("entity-info-card", {
@@ -70,15 +69,6 @@ const EntityInfoCard = (
       ...mergedProps.sections,
     ];
   }, [mergedProps]);
-
-  const hasState = Boolean(props.state);
-  const hasAccountBalance = Boolean(props.accountBalance);
-
-  if (hasState && hasAccountBalance) {
-    logger.error(
-      "You provided both state and accountBalance, only state will render"
-    );
-  }
 
   return (
     <EntityCardContainer testId={testId} overrideTestId fill>