+++ src/components/Icon/Icon.component.tsx
@@ -1,6 +1,7 @@
-import { useMemo, lazy, Suspense } from "react";
+import { useMemo, Suspense, type LazyExoticComponent } from "react";
 import styled from "styled-components";
 
+import * as lazyIcons from "./__generated__/index.lazy.ts";
 import { iconFallbackStyles } from "./Icon.styles.ts";
 import {
   DefaultIconProps,
@@ -9,11 +10,11 @@ import {
   IconProps,
   type MissingIcon,
 } from "./Icon.types";
-import { isMissingIcon, loadIcon, useIconSize } from "./Icon.utils.tsx";
+import { isGeneratedIcon, useIconSize } from "./Icon.utils.tsx";
 
 import { mergeDefaultProps, useLogger } from "@utils";
 
-import type { LazyExoticComponent } from "react";
+import type { Nullable } from "@ubloimmo/front-util";
 
 const defaultIconProps: DefaultIconProps = {
   size: "s-4",
@@ -26,17 +27,13 @@ const defaultIconProps: DefaultIconProps = {
  *
  * @remarks Determines the icon name based on the provided name, then looks up the corresponding icon component.
  *
- * @version 0.0.2
+ * @version 0.0.3
  *
  * @param {IconProps} props - The props for the icon component.
  * @return {JSX.Element | null} The rendered icon component or null if the icon component is not found.
  */
 const Icon = (props: IconProps) => {
   const { warn } = useLogger("Icon");
-  // runtime variable is needed instead of state or ref
-  // in order to be updated during lazy load
-  // only lasts 1 render
-  let iconIsMissing = false;
 
   if (!props.name) warn("Missing name prop");
   const { name, color, size } = useMemo(
@@ -45,30 +42,26 @@ const Icon = (props: IconProps) => {
   );
 
   const IconComponent = useMemo<
-    | LazyExoticComponent<GeneratedIcon | MissingIcon>
-    | GeneratedIcon
-    | MissingIcon
+    Nullable<
+      LazyExoticComponent<GeneratedIcon> | LazyExoticComponent<MissingIcon>
+    >
   >(() => {
-    return lazy(async () => {
-      const icon = await loadIcon(name, warn);
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-      iconIsMissing = isMissingIcon(icon.default);
-      return icon;
-    });
-  }, [name, warn]);
+    return lazyIcons[name] ?? null;
+  }, [name]);
 
   // sanitize size before passing it as svg width & height
   const parsedSize = useIconSize(size, warn);
 
-  if (!IconComponent || iconIsMissing) {
+  if (!isGeneratedIcon(IconComponent)) {
     warn(`No icon component found for name "${name}"`);
     return null;
   }
   return (
     <Suspense
+      key={`icon-suspense-${name}`}
       fallback={<IconFallback data-testid="icon-fallback" $size={parsedSize} />}
     >
-      <IconComponent size={parsedSize} color={color} />
+      <IconComponent key={`icon-${name}`} size={parsedSize} color={color} />
     </Suspense>
   );
 };