+++ CHANGELOG.md
@@ -7,12 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
 - `v0.0.7` InputLabel component
   - Decrease text size from `m` to `s`
 
 - `v0.0.4` InputAssistiveText component
   - Decrease text size from `s` to `xs`
   - Add `assistiveTextIcon` property to render an icon next to the assistive text
+
+- `v0.0.3` Icon component
+  - Add missing keys to stabilize
+  - Fix bug where lazy loading specific icons would make them flash upon parent's state update
+  - Update generation script to generate `index.lazy.ts` file
+
+- `v0.0.7` Tooltip component
+  - Change RenderedChildren rendering from `useCallback` to `useMemo` for better stability and less re-renders
+
+- `v0.0.8` Badge component
+  - Hide overflowing badge label with ellipsis
+
+- `v0.0.7` Chip component
+  - Set max-width to 100% to prevent overflow
+
+- `v0.0.13` EntityInfoCard component
+  - Remove error log when rendering both `state` and `accountBalance` props, since default props contain state
+
+- `v0.0.3` ListFilter component
+  - Fix overflow when rendering a single selected option's chip whith a long label
+  - Set scrollbar width to `thin`
+
+- `v0.0.9` Field component
+  - Update `useFieldValidity` hook to use `useUikitTranslation` for error messages
 
 ## 1.4.3 - 2025-02-03
 ### Fixed